Fix c source paths to be relative to documentation root
The logic for building up the c module name utilizes the sphinx
environment method `relfn2path`.  This method will normally build up the
path to be relative to the current document.  If one prefixes the
filename with `/` or a path separator then the path will be relative to
the documentation source root directory.  This fix forces the leading
'/` into the source paths when searching for c modules.

Fixes: #1